+++ CyRK/cy/common.pyx
@@ -2,6 +2,7 @@
 # cython: boundscheck=False, wraparound=False, nonecheck=False, cdivision=True, initializedcheck=False
 import cython
 
+from libc.math cimport fmax, fmin
 from libc.math cimport INFINITY as INF
 from libc.float cimport DBL_EPSILON as EPS
 from libc.stdint cimport SIZE_MAX, INT32_MAX
@@ -18,8 +19,8 @@ cdef double MAX_FACTOR       = 10.  # Maximum allowed increase in a step size.
 cdef double MAX_STEP         = INF
 cdef double EPS_10           = EPS * 10.
 cdef double EPS_100          = EPS * 100.
-cdef Py_ssize_t MAX_SIZET_SIZE = <Py_ssize_t>(0.95 * SIZE_MAX)
-cdef Py_ssize_t MAX_INT_SIZE   = <Py_ssize_t>(0.95 * INT32_MAX)
+cdef size_t MAX_SIZET_SIZE = <size_t>(0.95 * SIZE_MAX)
+cdef size_t MAX_INT_SIZE   = <size_t>(0.95 * INT32_MAX)
 
 # # Memory management constants
 # Assume that a cpu has a L1 of 300KB. Say that this progam will have access to 75% of that total.
@@ -37,15 +38,15 @@ cdef void interpolate(
         double* time_domain_reduced,
         double_numeric* target_array_full,
         double_numeric* target_array_reduced,
-        Py_ssize_t t_len_full,
-        Py_ssize_t t_len_reduced,
-        Py_ssize_t target_len,
+        size_t t_len_full,
+        size_t t_len_reduced,
+        size_t target_len,
         bool_cpp_t is_complex
         ) noexcept:
     """ Interpolate the results of a successful integration over the user provided time domain, `time_domain_full`. """
 
     # Setup loop variables
-    cdef Py_ssize_t i, j
+    cdef size_t i, j
 
     # Build a pointer array that will contain only 1 y for all ts in time_domain_full
     cdef double_numeric* array_slice_ptr
@@ -96,3 +97,40 @@ cdef void interpolate(
     # Release memory of any temporary variables
     PyMem_Free(array_slice_ptr)
     PyMem_Free(interpolated_array_slice_ptr)
+
+cdef size_t find_expected_size(
+        size_t y_size,
+        size_t num_extra,
+        double t_delta_abs,
+        double rtol_min,
+        bool_cpp_t capture_extra,
+        bool_cpp_t is_complex) noexcept nogil:
+
+    cdef double temp_expected_size
+    # Pick starting value that works with most problems
+    temp_expected_size = 500.0
+    # If t_delta_abs is very large or rtol is very small, then we may need more. 
+    temp_expected_size = \
+        fmax(
+            temp_expected_size,
+            fmax(
+                t_delta_abs / ARRAY_PREALLOC_TABS_SCALE,
+                ARRAY_PREALLOC_RTOL_SCALE / rtol_min
+                )
+            )
+    # Fix values that are very small/large
+    temp_expected_size = fmax(temp_expected_size, MIN_ARRAY_PREALLOCATE_SIZE)
+
+    if is_complex:
+        max_expected = MAX_ARRAY_PREALLOCATE_SIZE_DBL
+    else:
+        max_expected = MAX_ARRAY_PREALLOCATE_SIZE_DBLCMPLX
+    if capture_extra:
+        max_expected /= (y_size + num_extra)
+    else:
+        max_expected /= y_size
+    
+    temp_expected_size = fmin(temp_expected_size, max_expected)
+    # Store result as int
+    cdef size_t expected_size_to_use = <size_t>temp_expected_size
+    return expected_size_to_use